% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSNEgis-package.R
\docType{package}
\name{LAGOSNEgis-package}
\alias{LAGOSNEgis}
\alias{LAGOSNEgis-package}
\title{LAGOSNEgis: Extra functions to interact with the GIS module of LAGOSNE}
\description{
Extra Functions to Supplement LAGOSNE.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cont-limno/LAGOSNEgis}
  \item Report bugs at \url{https://github.com/cont-limno/LAGOSNEgis/issues}
}

}
\author{
\strong{Maintainer}: Joseph Stachelek \email{stachel2@msu.edu}

}
\keyword{internal}
