% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{lagosnegis_get}
\alias{lagosnegis_get}
\title{Get LAGOSNEgis data}
\usage{
lagosnegis_get(
  dest_folder = lagosnegis_dir(),
  overwrite = FALSE,
  cleanup = TRUE
)
}
\arguments{
\item{dest_folder}{file.path defaults to lagosnegis_path()}

\item{overwrite}{logical}

\item{cleanup}{logical. Delete intermediate file downloads?}
}
\description{
Get data from the GIS module of LAGOSNE
}
\examples{
\dontrun{
lagosnegis_get()
}
}
\references{
Soranno P., K. Cheruvelil. 2017. LAGOS-NE-GIS v1.0: A module for
 LAGOS-NE, a multi-scaled geospatial and temporal database of lake
 ecological context and water quality for thousands of U.S. Lakes:
 2013-1925. Environmental Data Initiative.
 http://dx.doi.org/10.6073/pasta/fb4f5687339bec467ce0ed1ea0b5f0ca.
 Dataset accessed 9/26/2017.
}
