% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lagosnegis_path}
\alias{lagosnegis_path}
\title{Get the path to LAGOSNEgis data}
\usage{
lagosnegis_path()
}
\description{
Get the path to LAGOSNEgis data
}
\details{
Can be overridden with the LAGOSNEGIS_PATH environment variable
}
