% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{load_xwalk}
\alias{load_xwalk}
\title{Load cross walk table between LAGOSNE and NHD}
\usage{
load_xwalk()
}
\description{
Load cross walk table between LAGOSNE and NHD
}
\examples{
\dontrun{
xwalk <- load_xwalk()
}
}
