% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_gis}
\alias{query_gis}
\title{Query LAGOS GIS}
\usage{
query_gis(layer, id_name, ids, crs = albers_conic(),
  gis_path = gis_path_default())
}
\arguments{
\item{layer}{character layer name}

\item{id_name}{selection column}

\item{ids}{feature ids to select}

\item{crs}{character projection info defaults to gis_path_default()}

\item{gis_path}{character path to LAGOSNE GIS gpkg}
}
\description{
Query LAGOS GIS
}
\examples{
\dontrun{

library(sf)
st_layers(gis_path_default())

library(gdalUtils)
ogrinfo(gis_path_default(), "HU12", so = TRUE)

library(mapview)
res_iws <- query_gis("IWS", "lagoslakeid", c(34352))
res_lake <- query_gis("LAGOS_NE_All_Lakes_4ha", "lagoslakeid", 34352)
res_pnt <- query_gis("LAGOS_NE_All_Lakes_4ha_POINTS", "lagoslakeid", 34352)
mapview(res_iws) + mapview(res_lake) + mapview(res_pnt)

res <- query_gis("IWS", "lagoslakeid", c(7010))
res <- query_gis("HU12", "ZoneID", c("HU12_1"))
}
}
