% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_gis_}
\alias{query_gis_}
\title{Raw (non-vectorized) query of LAGOS GIS}
\usage{
query_gis_(gis_path = lagosnegis_path(), query, crs = albers_conic())
}
\arguments{
\item{gis_path}{file path}

\item{query}{SQL string}

\item{crs}{coordinate reference system string or epsg code}
}
\description{
Raw (non-vectorized) query of LAGOS GIS
}
\examples{
\dontrun{
res <- query_gis_(
   gis_path = LAGOSNEgis:::lagosnegis_path(),
   query = "SELECT * FROM IWS WHERE lagoslakeid IN ('7010');",
   crs = LAGOSextra:::albers_conic())

# query nested hucs
hu4 <- query_gis("HU4", "ZoneID", c("HU4_5"))
hu8s <- query_gis_(
    gis_path = LAGOSNEgis:::lagosnegis_path(),
    query = "SELECT * FROM HU8 WHERE HUC8 LIKE '0415\%';",
    crs = LAGOSextra:::albers_conic())

}

}
