% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_wbd}
\alias{query_wbd}
\title{Query watershed boundary for LAGOS lakes}
\usage{
query_wbd(lagoslakeid, gis_path = gis_path_default(), crs = albers_conic(),
  utm = TRUE)
}
\arguments{
\item{lagoslakeid}{numeric}

\item{gis_path}{file.path to LAGOSNE GIS gpkg}

\item{crs}{projection string or epsg code}

\item{utm}{logical convert crs to utm}
}
\description{
Query watershed boundary for LAGOS lakes
}
\examples{
\dontrun{
library(mapview)
res <- query_wbd(lagoslakeid = c(5371))
mapview(res)

res <- query_wbd(lagoslakeid = c(2057, 3866, 1500, 3386, 2226, 1637, 6874, 7032, 1935, 6970, 5331))

}
}
