% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_gis}
\alias{query_gis}
\title{Query LAGOS GIS}
\usage{
query_gis(gis_path = NA, query, crs)
}
\arguments{
\item{gis_path}{file path}

\item{query}{SQL string}

\item{crs}{coordinate reference system string or epsg code}
}
\description{
Query LAGOS GIS
}
