% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_gis}
\alias{query_gis}
\title{Query LAGOS GIS}
\usage{
query_gis(layer, id_name = NULL, ids = NULL, crs = albers_conic(),
  gis_path = lagosnegis_path())
}
\arguments{
\item{layer}{character layer name}

\item{id_name}{selection column}

\item{ids}{feature ids to select}

\item{crs}{character projection info defaults to lagosnegis_path()}

\item{gis_path}{character path to LAGOSNE GIS gpkg}
}
\description{
Query LAGOS GIS
}
\examples{
\dontrun{

library(sf)
st_layers(lagosnegis_path())

library(gdalUtils)
ogrinfo(lagosnegis_path(), "HU12", so = TRUE)

states   <- query_gis("STATE")

library(mapview)
state    <- query_gis("STATE", "State_Name", "Michigan")
res_iws  <- query_gis("IWS", "lagoslakeid", c(34352))
res_lake <- query_gis("LAGOS_NE_All_Lakes_4ha", "lagoslakeid", 34352)
res_pnt  <- query_gis("LAGOS_NE_All_Lakes_4ha_POINTS", "lagoslakeid", 34352)
mapview(state) + mapview(res_iws) + mapview(res_lake) + mapview(res_pnt)

res <- query_gis("IWS", "lagoslakeid", c(7010))
res <- query_gis("HU12", "ZoneID", c("HU12_1"))
res <- query_gis("HU8", "ZoneID", c("HU8_100"))
res <- query_gis("HU4", "ZoneID", c("HU4_5"))

# query mutiple feature ids
res <- query_gis("IWS", "lagoslakeid", c(7010, 1))
}
}
