% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_gis}
\alias{query_gis}
\title{Query LAGOS GIS}
\usage{
query_gis(layer, id_name, ids, crs = albers_conic(),
  gis_path = gis_path_default())
}
\arguments{
\item{layer}{character layer name}

\item{id_name}{selection column}

\item{ids}{feature ids to select}

\item{crs}{character projection info defaults to gis_path_default()}

\item{gis_path}{character path to LAGOSNE GIS gpkg}
}
\description{
Query LAGOS GIS
}
