% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_gis_}
\alias{query_gis_}
\title{Raw (non-vectorized) query of LAGOS GIS}
\usage{
query_gis_(gis_path, query, crs)
}
\arguments{
\item{gis_path}{file path}

\item{query}{SQL string}

\item{crs}{coordinate reference system string or epsg code}
}
\description{
Raw (non-vectorized) query of LAGOS GIS
}
\examples{
\dontrun{
res <- query_gis_(gis_path = "/home/jose/.local/share/LAGOS-GIS/lagos-ne_gis.gpkg",
query = "SELECT * FROM IWS WHERE lagoslakeid IN ('7010');",
crs = albers_conic())
}

}
