% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_wbd}
\alias{query_wbd}
\title{Query watershed boundary for LAGOS lakes}
\usage{
query_wbd(lagoslakeid, crs = NA)
}
\arguments{
\item{lagoslakeid}{numeric}

\item{crs}{projection string or epsg code}
}
\description{
Query watershed boundary for LAGOS lakes
}
\examples{
\dontrun{
res <- query_wbd(lagoslakeid = 7429)
}
}
