% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_wbd}
\alias{query_wbd}
\title{Query watershed boundary for LAGOS lakes}
\usage{
query_wbd(lagoslakeid, gis_path = NA, crs = NA)
}
\arguments{
\item{lagoslakeid}{numeric}

\item{gis_path}{file.path to LAGOSNE GIS gpkg}

\item{crs}{projection string or epsg code}
}
\description{
Query watershed boundary for LAGOS lakes
}
\examples{
\dontrun{
library(sf)
res <- query_wbd(lagoslakeid = c(5371, 4559))
plot(res)
}
}
